requirements.txt
# Core Framework
streamlit==1.51.0
python-dotenv==1.2.1

# Spotify Integration
spotipy==2.25.1

# LangChain & AI
langchain==1.0.3
langchain-google-genai==3.0.0
langchain-chroma==1.0.0
langchain-community==0.4.1
langchain-huggingface==1.0.1
langgraph==1.0.2

# Vector Database & Embeddings
chromadb==1.3.0
sentence-transformers==5.1.2

# Google AI
google-generativeai==0.3.2

# Data Processing
pandas==2.3.3
numpy==2.3.4

# Web Scraping (si lo necesitas)
beautifulsoup4==4.14.2

# HTTP Requests
requests==2.32.5
httpx==0.28.1